Added new function for creating virtual pageviews
Sometimes it’s useful to log something as a pageview rather than an event (if you’re making a funnel for a goal for instance). This script allows you to use data-attributes to make this happen, at the moment it only works for on form submit events. It can scrape your form a pick up values and post them as values to Google too.

Use `data-virtual-pageview="page/slug/name"` to trigger the script to listen for form submissions.

Updated readme with example